# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sales operations API. All routes except login require an authenticated employee session.
servers:
  - url: /api
paths:
...
  /auth/login:
    post:
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/LoginInput"}
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Employee"}
  /auth/mobile-login:
    post:
      operationId: mobileLogin
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/LoginInput"}
      responses:
        "200":
          description: Native app session
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MobileSession"}
...
  /auth/password:
    post:
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/PasswordInput"}
      responses:
        "200":
          description: Changed
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Message"}
...
  /imports/preview:
    post:
      operationId: previewImport
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: {$ref: "#/components/schemas/ImportUpload"}
      responses:
        "200":
          description: Validated preview
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ImportPreview"}
  /imports/confirm:
    post:
      operationId: confirmImport
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ImportConfirmation"}
      responses:
        "200":
          description: Applied atomically
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ImportResult"}
...
  /exports/{kind}:
    get:
      operationId: exportData
      parameters:
        - name: kind
          in: path
          required: true
          schema: {type: string, enum: [users, customers, products, branches, sales, attendance]}
      responses:
        "200":
          description: Authorized Excel workbook
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema: {type: string, format: binary}
components:
...
  schemas:
...
    Message:
      type: object
      required: [message]
      properties:
        message: {type: string}
    MobileSession:
      type: object
      required: [user, token]
      properties:
        user: {$ref: "#/components/schemas/Employee"}
        token: {type: string}
    LoginInput:
      type: object
      required: [employeeId, password]
      properties:
        employeeId: {type: string}
        password: {type: string}
    PasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: {type: string}
        newPassword: {type: string, minLength: 12}
    Employee:
      type: object
      required: [id, employeeId, nameAr, role, active, mustChangePassword, effectivePermissions]
      properties:
        id: {type: integer}
        employeeId: {type: string}
        nameAr: {type: string}
        nameEn: {type: ["string", "null"]}
        role: {type: string}
        active: {type: boolean}
        mustChangePassword: {type: boolean}
        areaId: {type: ["integer", "null"]}
        branchId: {type: ["integer", "null"]}
        permissions:
          type: ["array", "null"]
          items: {$ref: "#/components/schemas/Permission"}
          description: Null uses all permissions allowed by the role; an array narrows them.
        effectivePermissions:
          type: array
          items: {$ref: "#/components/schemas/Permission"}
    Permission:
      type: string
      enum: [dashboard.view, users.view, users.manage, customers.view, customers.manage, branches.view, branches.manage, products.view, products.manage, imports.manage, exports.view, sales.view, sales.write, attendance.view, attendance.write]
...
    ImportUpload:
      type: object
      required: [kind, file]
      properties:
        kind: {type: string, enum: [products, users]}
        file: {type: string, format: binary}
    ImportRow:
      type: object
      required: [code, name, change]
      properties:
        code: {type: string}
        name: {type: string}
        price: {type: ["string", "null"]}
        change: {type: string}
    ImportPreview:
      type: object
      required: [token, sheet, total, newCount, updateCount, rows]
      properties:
        token: {type: string}
        sheet: {type: string}
        total: {type: integer}
        newCount: {type: integer}
        updateCount: {type: integer}
        rows: {type: array, items: {$ref: "#/components/schemas/ImportRow"}}
    ImportConfirmation:
      type: object
      required: [token]
      properties:
        token: {type: string}
    ImportResult:
      type: object
      required: [created, updated]
      properties:
        created: {type: integer}
        updated: {type: integer}
...