# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Sales operations API. All routes except login require an authenticated employee session.
 * OpenAPI spec version: 0.1.0
 */
import type { Permission } from './permission';

export interface Employee {
  id: number;
  employeeId: string;
  nameAr: string;
  /** @nullable */
  nameEn?: string | null;
  role: string;
  active: boolean;
  mustChangePassword: boolean;
  /** @nullable */
  areaId?: number | null;
  /** @nullable */
  branchId?: number | null;
  /**
     * Null uses all permissions allowed by the role; an array narrows them.
     * @nullable
     */
  permissions?: Permission[] | null;
  effectivePermissions: Permission[];
}