# lib/api-zod/src/generated/types/importRow.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Sales operations API. All routes except login require an authenticated employee session.
 * OpenAPI spec version: 0.1.0
 */

export interface ImportRow {
  code: string;
  name: string;
  /** @nullable */
  price?: string | null;
  change: string;
}